requirements.txt
flask
torch
torchvision
ultralytics
opencv-python-headless
gdown
numpy<2
pillow
